Fortify
Former-commit-id: dc0d1728ac41f23bc8c58e48576798ad06205e1e

+++ met/src/libcode/vx_pb_util/do_blocking.cc
@@ -118,7 +118,7 @@ switch ( padsize )  {
    case padsize_4:
       bytes = 4;
       b = (unsigned char *) (&I);
-      I = value;
+      I = (unsigned int) value;
       break;
 
    case padsize_8:
